Move non-plugin modules to other folders
