+++ src/analyzer_nlp/load_scancode_data.py
